% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{canonical_correlation_analysis}
\alias{canonical_correlation_analysis}
\title{Canonical Correlation Analysis}
\usage{
canonical_correlation_analysis(X, Y, epsilon = 1e-04)
}
\arguments{
\item{X}{numeric matrix (n * p) with n observations of p variables}

\item{Y}{numeric matrix (n * q) with n observations of q variables}

\item{epsilon}{numeric scalar tolerance for rank reduction of the
input matrices}
}
\value{
A list containing the following components
        \item{xcoef}{Estimated coefficients for the X variables.}
        \item{ycoef}{Estimated coefficients for the Y variables.}
        \item{cor}{Matrix of correlations.}
}
\description{
CCA will find pairs of vectors \eqn{(w,v)} such that projections \eqn{Xw} and \eqn{Yv} have
maximal possible correlations (the pairs will be ordered in the order of
decreasing correlations). Projection vectors are normalized such that the
variance of \eqn{Xw} and of \eqn{Yv} is equal to \eqn{1}. This means that projections are
not only correlated, but "on the same scale" and hence can be directly
compared.
}
\examples{
library(MASS)
library(pracma)
X = mvrnorm(1000, c(0,0), eye(2))
cca = canonical_correlation_analysis(X, X)

}

