% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{canonical_correlation_forest.default}
\alias{canonical_correlation_forest.default}
\title{Defaultf S3 method.}
\usage{
\method{canonical_correlation_forest}{default}(X, Y_one_hot,
  nr_of_trees = 200)
}
\arguments{
\item{X}{numeric matrix (n * p) with n observations of p variables}

\item{Y_one_hot}{muneric matrix with n observations of q variables
(one_hot_encoded)}

\item{nr_of_trees}{nr of trees the forest will be composed of}
}
\value{
returns an object of class "canonical_correlation_forest",
where an object of class "canonical_correlation_forest" is a list
containing the following components:
   X,Y_one_hot The original input data
   forest a vector of length nr_of_trees with objects of class
   "canonical_correlation_tree"
}
\description{
Defaultf S3 method.
}

