% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{canonical_correlation_forest}
\alias{canonical_correlation_forest}
\title{Canonical Correlation Forest}
\usage{
canonical_correlation_forest(x, ...)
}
\arguments{
\item{x}{Either a formula or a matrix.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function for creating canonical correlation forests.
}

