% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cct.R
\name{canonical_correlation_tree}
\alias{canonical_correlation_tree}
\title{Creates a canonical correlation tree.}
\usage{
canonical_correlation_tree(X, Y, depth = 0, options = list(minPointsForSplit
  = 2, maxDepthsplit = Inf))
}
\arguments{
\item{X}{numeric matrix (n * p) wiht n observations of p variables}

\item{Y}{muneric matrix (n * q) wiht n observations of q variables}

\item{depth}{The subtree depth}

\item{options}{list containing the following options for the tree
construction:
     minPointsForSplit if the number of datapoints is smaller than
       this value, a leaf is constructed
     maxDepthSplit if the current depth is grater than
       this value, a leaf is constructed}
}
\value{
returns an object of class "canonical_correlation_tree",
where an object of class  "canonical_correlation_tree" is a list
containing at the following components:
   isLeaf boolean wether the tree is a leaf
   trainingCounts number of training examples this tree got (ncol(X))
   iIn feature indices that the node got, needed for prediction
   decisionProjection numeric matrix The projection matrix that was
     used to find the best split point
   lessthanChild Reference to the left subtree
   greaterthanChild Reference to the right subtree
}
\description{
Creates a canonical correlation tree.
}

