% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_missclassification_rate}
\alias{get_missclassification_rate}
\title{Helper function to print prediction accuracy for a model.}
\usage{
get_missclassification_rate(model, data_test, ...)
}
\arguments{
\item{model}{a model object for which prediction is desired. E.g. object of class
\code{canonical_correlation_forrest}, \code{canonical_correlation_tree},
\code{tree} or \code{randomForest}.}

\item{data_test}{A data frame or a matrix containing the test data.}

\item{...}{Further arguments passed to model.predict()}
}
\description{
Helper function to print prediction accuracy for a model.
}
