% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{plot.canonical_correlation_forest}
\alias{plot.canonical_correlation_forest}
\title{Visualization of canonical correlation forest}
\usage{
\method{plot}{canonical_correlation_forest}(...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}
}
\description{
TODO: document
}

