% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cct.R
\name{plot.canonical_correlation_tree}
\alias{plot.canonical_correlation_tree}
\title{Decision surface for a Canonical Correlation Tree}
\usage{
\method{plot}{canonical_correlation_tree}(x, dataX, dataY, ...)
}
\arguments{
\item{x}{fitted model of class canonical_correlation_tree}

\item{dataX}{numeric matrix (n * p) with n observations of p variables.
Note: n and p have to match the dimensions of the training data.}

\item{dataY}{classes of the n observations}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plots the decision surface for a CCT, training dataa has to be 2d.
}

