% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_decision_surface}
\alias{plot_decision_surface}
\title{Helper function to plot classifier decision surface.}
\usage{
plot_decision_surface(model, X, Y, title = NULL, interpolate = FALSE, ...)
}
\arguments{
\item{model}{a model object for which prediction is desired. E.g. object of class
\code{canonical_correlation_forrest}, \code{canonical_correlation_tree},
\code{tree} or \code{randomForest}.}

\item{X}{Numeric matrix (n * 2) with n observations of 2 variables}

\item{Y}{Numeric matrix with n observations of 1 variable}

\item{title}{Title text for the plot.}

\item{interpolate}{If TRUE interpolate linearly, if FALSE (the default)
don't interpolate.}

\item{...}{Further arguments passed to model.predict()}
}
\description{
This function generates a plot (ggplot2) of the decision surface for a 2d classifier.
}

