% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{predict.canonical_correlation_forest}
\alias{predict.canonical_correlation_forest}
\title{Predictions from a Canonical Correlation Forest}
\usage{
\method{predict}{canonical_correlation_forest}(object, newData,
  options = list(verbose = F), ...)
}
\arguments{
\item{object}{fitted model of class canonical_correlation_forest}

\item{newData}{numeric matrix (n * p) with n observations of p variables.
Note: n and p have to match the dimensions of the training data.}

\item{options}{list of options}

\item{...}{further arguments passed to or from other methods.}
}
\description{
## S3 method for class 'canonical_correlation_forest'
}

