% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cct.R
\name{predict.canonical_correlation_tree}
\alias{predict.canonical_correlation_tree}
\title{Predictions from a Canonical Correlation Tree}
\usage{
\method{predict}{canonical_correlation_tree}(object, newData, ...)
}
\arguments{
\item{object}{fitted model of class canonical_correlation_tree}

\item{newData}{numeric matrix (n * p) with n observations of p variables.
Note: n and p have to match the dimensions of the training data.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
## S3 method for class 'canonical_correlation_tree'
}

