% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spirals}
\alias{spirals}
\title{Spiral dataset}
\format{
A data frame with 10000 rows and 3 variables:
\describe{
  \item{x}{numeric scalar: x-coordinate}
  \item{y}{numeric scalar: y-coordinate}
  \item{class}{integer: either 1,2 or 3}
}
}
\source{
Created by T. Rainforth, URL:
\url{https://bitbucket.org/twgr/ccf/raw/49d5fce6fc006bc9a8949c7149fc9524535ce418/Datasets/spirals.csv}
}
\usage{
data(spirals)
}
\description{
A dataset containing 3 interwinding spirals.
}
\keyword{datasets}
