% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cct.R
\name{canonical_correlation_tree}
\alias{canonical_correlation_tree}
\title{Computes a canonical correlation tree}
\usage{
canonical_correlation_tree(X, Y, depth = 0, minPointsForSplit = 2,
  maxDepthSplit = Inf, ancestralProbs = NULL)
}
\arguments{
\item{X}{Predictor matrix of size \eqn{n \times p} with \eqn{n} observations and \eqn{p}
variables.}

\item{Y}{Predicted values as a matrix of size \eqn{n \times p} with \eqn{n} observations
and \eqn{p} variables.}

\item{depth}{Depth of subtree.}

\item{minPointsForSplit}{Optional parameter setting the threshold when to construct a
leaf (default: 2). If the number of data points is smaller than this value, a leaf is
constructed.}

\item{maxDepthSplit}{Optional parameter controlling the construction of leaves after a
certain depth (default: \code{Inf}). If the current depth is greater than this value,
a leaf is constructed.}

\item{ancestralProbs}{Probabilities of ancestors. Defualt is \code{NULL} as these are
then calculated automatically.}
}
\value{
Function returns an object of class \code{canonical_correlation_tree},
where the object is a list containing at the following components:
\itemize{
  \item{isLeaf}{Boolean whether the tree is a leaf itself.}
  \item{trainingCounts}{Number of training examples for constructing this tree (i.e.
  number of rows in input argument \code{X}).}
  \item{indicesFeatures}{Feature indices which the node received, as needed for
  prediction.}
  \item{decisionProjection}{Numeric matrix containing the projection matrix that was
     used to find the best split point.}
  \item{refLeftChild}{Reference to the left subtree.}
  \item{refRightChild}{Reference to the right subtree.}
}
}
\description{
This function computes a single canonical correlation tree given its input values.
}

