% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\docType{data}
\name{spirals}
\alias{spirals}
\title{An artificial spiral dataset}
\format{A data frame with 10000 rows and 3 variables}
\description{
A dataset containing 3 interwinding spirals. The variables of the
10,000 datapoints are:
}
\details{
\itemize{
  \item x numeric scalar x-coordinate
  \item y numeric scalar y-coordinate
  \item class integer either 1,2 or 3
}
}
\references{
\url{https://bitbucket.org/twgr/ccf/raw/49d5fce6fc006bc9a8949c7149fc9524535ce418/Datasets/spirals.csv}
}
\keyword{datasets}

